qmake: Airmap: Download SDK on demand
Since the build process is cut from the internet once the build process is starting in flatpak's builder, we need to download the SDK before running our build there.
Therefore I had to expand qmake with the possibility to do the following:
* Adding a variable which allows to specify the location of the downloaded SDK
* If the file already exists at the location specified, skip the download.
* If the file has been downloaded, ensure it gets removed afterwards just like qmake was doing before.

Another difference is that I'm not using Wget, but Curl to do the download now. That's again caused by the Flatpak infrastructure, which only provides Curl in the Freedesktop SDK. I could build my own Wget there, which I use for the build process only, but better took the time and quickly rewrote the commands to be more efficient.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com>